Backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
sqlalchemy
pydantic
email-validator
